Choose instances based on available space
I removed volumes for containers as the data gets preserved, and in the
original version of the task it doesn't seem important to preserve any
data between restarts. Also it seems when `scale` is used, the volume
gets shared between all containers and you cannot easily assign separate
volumes to separate machines in a scale group.