Add DataDistributor and rollback on save (again)
Now I seem to like it a bit more. ChunkMaster is separate from storage-
related activities and simply creates chunks based on some internal
logic and provided info about known storages.

DataDistributor grew from ChunkMaster as I didn't like the area of
responsibility of the latter, and it is now managing the information
about storages and does data distribution/restoration logic.

All of that mainly was done to put transaction rollback into the place
which seems "correct".

Tests are in a bad shape, but I'll fix them shortly